Feat: new files js and css
It was create 1 file css named modal css, and was modificaste 2 files such as; footer.css and index.css. In the other hand was create 4 file js which are; index.js , node.js, printCard.js, printModal.js. Also was create a new folder named utils which has a new file named data.js this file will be the data-base for the page. Finally the last thing was add new images and icons for the data base