Feat: new file and new features
It was created new file named useSplide for hooks for this was added some code in the files; jsconfig.paths.json and configoverrides.js to set alias path for hooks folder. Also was added new features in the files; Landing.jsx and SpliderGrid.jsx. The last thing was added new styles in the scss files; index.scss and SliderGrid.scss