Feat: new features
It was fixed some bugs in the component Footer.jsx and was implemented new features in it. In the App.jsx was implemented the Header.jsx and Footer.jsx. It was removed the Header.jsx and Footer.jsx from Landing.jsx. The last thing was added new data in the file static.data.js