Style: restructuring css files
It was created a new scss file in the new folder called Scss which in it were added the scss variables of the project. In the other hand the scss files were restructured as; Index.scss, App.scss. The last thing was delete de App.css files from the project that does not have any classes