Feat: new features, components and assets
It was created a new components which are; CardProductFavorite.jsx which was implemented in FavoriteList.jsx component, CardProductShopping.jsx which was implemented in MyShopping.jsx component, EmptyList.jsx which was implemented in FavoriteList.jsx and MyShopping.jsx component, EmptyShoppingCar.jsx which was implemented in ShoppingDrawer.jsx. Also were added new images to the project and implemented in the components EmptyList.jsx  and EmptyShoppingCar.jsx. The last thing was added new static data to the file static.data.js